fix(games,photo): BigInteger for scores, anti-cheat cap, better error logging
- game_scores.score: Integer → BigInteger (fixes int32 overflow crash)
- Cap score at 100,000 (anti-cheat for snake)
- Log OpenRouter error response body for debugging nano-banana failures
- Detect safety/blocked responses and return 400 instead of 502

+++ backend/app/models/game_score.py
@@ -14,6 +14,6 @@ class GameScore(Base):
     id: Mapped[str] = mapped_column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     user_tg_id: Mapped[int] = mapped_column(BigInteger, index=True, nullable=False)
     game: Mapped[str] = mapped_column(String(20), nullable=False, default="snake")
-    score: Mapped[int] = mapped_column(Integer, nullable=False)
+    score: Mapped[int] = mapped_column(BigInteger, nullable=False)
     month: Mapped[str] = mapped_column(String(7), nullable=False)  # YYYY-MM
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
